refactor: finalize dynamic enum structure per review feedback
- Renamed PriceRepresentationMode to PriceRepresentationModeDefinition.
- Reorganized price representation mode components into the dataaccess.channel.pricerepresentationmode package.
- Updated all internal references to maintain functionality.
