+++ service/src/main/java/io/commercestacksolutions/priceproviderservice/facade/channel/mapper/ChannelEntityMapper.java
@@ -5,9 +5,9 @@ import io.commercestacksolutions.commons.mapper.RestRequestMappingContext;
 import io.commercestacksolutions.commons.mapper.exception.DataMappingException;
 import io.commercestacksolutions.priceproviderservice.commons.messagekeys.MessageKeys;
 import io.commercestacksolutions.priceproviderservice.dataaccess.channel.entity.ChannelEntity;
+import io.commercestacksolutions.priceproviderservice.dataaccess.channel.pricerepresentationmode.PriceRepresentationModeType;
 import io.commercestacksolutions.priceproviderservice.dataaccess.country.entity.CountryEntity;
 import io.commercestacksolutions.priceproviderservice.facade.channel.restentity.ChannelRestEntity;
-import io.commercestacksolutions.priceproviderservice.service.publicprice.strategy.PriceRepresentationModeType;
 import io.commercestacksolutions.priceproviderservice.service.country.CountryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
